+++ src/command/scrub/executor.rs
@@ -113,11 +113,13 @@ impl ActionSink for Executor {
         info!("{action}");
 
         match action {
-            Action::MigrateNamespaceRegistry => {
-                self.metadata_store.migrate_namespace_registry().await?;
-            }
             Action::MigrateBlobIndex(digest) => {
                 self.metadata_store.migrate_blob_index(&digest).await?;
+            }
+            Action::PruneLegacyNamespaceRegistry => {
+                self.metadata_store
+                    .delete_legacy_namespace_registry()
+                    .await?;
             }
             Action::DeleteOrphanBlob(digest) => {
                 // Hold the `blob-data:{digest}` coarse lock (the same one